Deleted dead code and created controller.microservice package to hold microservice related controllers.
